+++ IBAnimatable/PresentSlideSegue.swift
@@ -7,7 +7,7 @@ import UIKit
 
 public class PresentSlideSegue: UIStoryboardSegue {
   public override func perform() {
-    destinationViewController.transitioningDelegate = PresenterManager.sharedManager().retrievePresenter(.Slide(direction: .Left, params: []))
+    destinationViewController.transitioningDelegate = PresenterManager.sharedManager().retrievePresenter(.Slide(toDirection: .Left, params: []))
     sourceViewController.presentViewController(destinationViewController, animated: true, completion: nil)
   }
 }