+++ IBAnimatable/SystemFlipAnimator.swift
@@ -16,29 +16,29 @@ public class SystemFlipAnimator: NSObject, AnimatedTransitioning {
   public var interactiveGestureType: InteractiveGestureType?
 
   // MARK: - private
-  private var fromDirection: TransitionFromDirection
+  private var fromDirection: TransitionDirection
   private var animationOption: UIViewAnimationOptions
   
-  public init(fromDirection: TransitionFromDirection, transitionDuration: Duration) {
+  public init(fromDirection: TransitionDirection, transitionDuration: Duration) {
     self.fromDirection = fromDirection
     self.transitionDuration = transitionDuration
     
     switch fromDirection {
     case .Right:
-      self.transitionAnimationType = .SystemFlip(direction: .Right)
-      self.reverseAnimationType = .SystemFlip(direction: .Left)
+      self.transitionAnimationType = .SystemFlip(fromDirection: .Right)
+      self.reverseAnimationType = .SystemFlip(fromDirection: .Left)
       self.animationOption = .TransitionFlipFromRight
     case .Top:
-      self.transitionAnimationType = .SystemFlip(direction: .Top)
-      self.reverseAnimationType = .SystemFlip(direction: .Bottom)
+      self.transitionAnimationType = .SystemFlip(fromDirection: .Top)
+      self.reverseAnimationType = .SystemFlip(fromDirection: .Bottom)
       self.animationOption = .TransitionFlipFromTop
     case .Bottom:
-      self.transitionAnimationType = .SystemFlip(direction: .Bottom)
-      self.reverseAnimationType = .SystemFlip(direction: .Top)
+      self.transitionAnimationType = .SystemFlip(fromDirection: .Bottom)
+      self.reverseAnimationType = .SystemFlip(fromDirection: .Top)
       self.animationOption = .TransitionFlipFromBottom
     default:
-      self.transitionAnimationType = .SystemFlip(direction: .Left)
-      self.reverseAnimationType = .SystemFlip(direction: .Right)
+      self.transitionAnimationType = .SystemFlip(fromDirection: .Left)
+      self.reverseAnimationType = .SystemFlip(fromDirection: .Right)
       self.animationOption = .TransitionFlipFromLeft
     }
     
@@ -52,6 +52,6 @@ extension SystemFlipAnimator: UIViewControllerAnimatedTransitioning {
   }
   
   public func animateTransition(transitionContext: UIViewControllerContextTransitioning) {
-    animateWithCATransition(transitionContext, type: SystemTransitionType.Flip, subtype: fromDirection.stringValue)
+    animateWithCATransition(transitionContext, type: SystemTransitionType.Flip, subtype: fromDirection.CATransitionSubtype)
   }
 }