+++ IBAnimatable/TransitionFromDirection.swift
@@ -5,17 +5,18 @@
 import UIKit
 
 /**
-TransitionFromDirection: convert from direction to CATransition Subtype used in `CATransition`
+TransitionDirection: used to specify the direction for the transition
 */
-public enum TransitionFromDirection {
+public enum TransitionDirection {
   case Left
   case Right
   case Top
   case Bottom
   case Forward
   case Backward
-  
-  var stringValue: String {
+
+  // Convert from direction to CATransition Subtype used in `CATransition`
+  var CATransitionSubtype: String {
     switch self {
     case .Left:
       return kCATransitionFromLeft